+++ hive/lib/src/hive_list.dart
@@ -2,33 +2,35 @@ part of hive;
 
 class HiveList<E extends HiveObject>
     with HiveCollectionMixin<E>, DelegatingListViewMixin<E> {
-  final HiveImpl _hive;
+  final Box Function(String) _getBox;
 
   final String _boxName;
 
-  final List<dynamic> _keys;
+  List<dynamic> _keys;
 
   Box _box;
 
-  HiveList(Box box)
-      : _hive = Hive as HiveImpl,
-        _box = box,
+  HiveList(Box box, {List<dynamic> keys})
+      : _getBox = null,
         _boxName = box.name,
-        _keys = [];
+        _box = box,
+        _keys = keys ?? [];
 
-  HiveList.fromKeys(String boxName, List<dynamic> keys)
-      : _hive = Hive as HiveImpl,
+  HiveList.internal(String boxName, List<dynamic> keys)
+      : _getBox = (Hive as HiveImpl).getBoxInternal,
         _boxName = boxName,
         _keys = keys;
 
-  HiveList.debug(String boxName, List<dynamic> keys, this._hive)
-      : _boxName = boxName,
+  @visibleForTesting
+  HiveList.debug(String boxName, List<dynamic> keys, HiveImpl hive)
+      : _getBox = hive.getBoxInternal,
+        _boxName = boxName,
         _keys = keys;
 
   @override
   Box get box {
     if (_box == null) {
-      var _box = _hive.getBoxInternal(_boxName);
+      _box = _getBox(_boxName);
       if (_box == null) {
         throw HiveError(
             'To use this list, you have to open the box "$_boxName" first.');
@@ -49,8 +51,18 @@ class HiveList<E extends HiveObject>
     return list;
   }
 
+  void _refreshKeys() {
+    var newKeys = <dynamic>[];
+    for (var key in _keys) {
+      if (box.containsKey(key)) {
+        newKeys.add(key);
+      }
+    }
+    _keys = newKeys;
+  }
+
   void _checkHiveObjectIsValid(E obj) {
-    if (obj.box == null || obj.box != box) {
+    if (obj.box != box) {
       throw HiveError('The HiveObject needs to be in the box "$_boxName".');
     }
   }
@@ -58,6 +70,7 @@ class HiveList<E extends HiveObject>
   @override
   void operator []=(int index, E value) {
     _checkHiveObjectIsValid(value);
+    _refreshKeys();
     _keys[index] = value.key;
   }
 
@@ -85,18 +98,21 @@ class HiveList<E extends HiveObject>
   @override
   void fillRange(int start, int end, [E fillValue]) {
     _checkHiveObjectIsValid(fillValue);
+    _refreshKeys();
     _keys.fillRange(start, end, fillValue.key);
   }
 
   @override
   set first(E value) {
     _checkHiveObjectIsValid(value);
+    _refreshKeys();
     _keys.first = value.key;
   }
 
   @override
   void insert(int index, E element) {
     _checkHiveObjectIsValid(element);
+    _refreshKeys();
     _keys.insert(index, element.key);
   }
 
@@ -105,6 +121,7 @@ class HiveList<E extends HiveObject>
     for (var element in iterable) {
       _checkHiveObjectIsValid(element);
     }
+    _refreshKeys();
     for (var element in iterable) {
       _keys.insert(index++, element);
     }
@@ -113,17 +130,19 @@ class HiveList<E extends HiveObject>
   @override
   set last(E value) {
     _checkHiveObjectIsValid(value);
+    _refreshKeys();
     _keys.last = value.key;
   }
 
   @override
   set length(int newLength) {
-    _keys.length;
+    _keys.length = newLength;
   }
 
   @override
   bool remove(Object value) {
     if (value is E) {
+      _checkHiveObjectIsValid(value);
       return _keys.remove(value.key);
     } else {
       return false;
@@ -132,43 +151,60 @@ class HiveList<E extends HiveObject>
 
   @override
   E removeAt(int index) {
+    _refreshKeys();
     var key = _keys.removeAt(index);
     return box.get(key) as E;
   }
 
   @override
   E removeLast() {
+    _refreshKeys();
     var key = _keys.removeLast();
     return box.get(key) as E;
   }
 
   @override
   void removeRange(int start, int end) {
+    _refreshKeys();
     _keys.removeRange(start, end);
   }
 
   @override
   void removeWhere(bool Function(E element) test) {
+    _refreshKeys();
     _keys.removeWhere((key) => test(box.get(key) as E));
   }
 
   @override
   void replaceRange(int start, int end, Iterable<E> replacement) {
-    _keys.replaceRange(start, end, replacement.map((e) => e.key));
+    for (var element in replacement) {
+      _checkHiveObjectIsValid(element);
+    }
+    _refreshKeys();
+    _keys.replaceRange(start, end, replacement.map<dynamic>((e) => e.key));
   }
 
   @override
   void retainWhere(bool Function(E element) test) {
+    _refreshKeys();
     _keys.retainWhere((key) => test(box.get(key) as E));
   }
 
   @override
   void setAll(int index, Iterable<E> iterable) {
+    for (var element in iterable) {
+      _checkHiveObjectIsValid(element);
+    }
+    _refreshKeys();
     _keys.setAll(index, iterable.map((e) => e.key));
   }
 
   @override
   void setRange(int start, int end, Iterable<E> iterable, [int skipCount = 0]) {
+    for (var element in iterable) {
+      _checkHiveObjectIsValid(element);
+    }
+    _refreshKeys();
     _keys.setRange(start, end, iterable.map((e) => e.key), skipCount);
   }
 
@@ -179,11 +215,12 @@ class HiveList<E extends HiveObject>
 
   @override
   void sort([int Function(E a, E b) compare]) {
+    _refreshKeys();
     _keys.sort((a, b) => compare(box.get(a) as E, box.get(b) as E));
   }
 
   HiveList<T> castHiveList<T extends HiveObject>() {
-    return HiveList<T>.fromKeys(_boxName, _keys);
+    return HiveList<T>.internal(_boxName, _keys);
   }
 
   @visibleForTesting